% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_international_name}
\alias{clean_international_name}
\title{Standardizes country and clubs for interntional competitions where only country is available using a dataset in the kickout library}
\usage{
clean_international_name(event)
}
\arguments{
\item{event}{event data}
}
\value{
standardized country and club names
}
\description{
Standardizes country and clubs for interntional competitions where only country is available using a dataset in the kickout library
}
\examples{
 \dontrun{
   clean_international_name(event)
}
}
\keyword{internal}
