% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_names}
\alias{clean_names}
\title{Standardizes names using a dataset in the kickout library.}
\usage{
clean_names(event)
}
\arguments{
\item{event}{A data frame containing event data with columns: title, surname, given_panel_name.}
}
\value{
A data frame with standardized names.
}
\description{
Standardizes names using a dataset in the kickout library.
}
\examples{
\dontrun{
clean_names(event)
}
}
\keyword{internal}
