% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_representing}
\alias{clean_representing}
\title{Standardizes club and country using a dataset in the kickout library}
\usage{
clean_representing(event)
}
\arguments{
\item{event}{Event data.}
}
\value{
A data frame with standardized country and club names.
}
\description{
Standardizes club and country using a dataset in the kickout library
}
\examples{
\dontrun{
clean_representing(event)
}
}
\keyword{internal}
