% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_competition_list.R
\name{create_competition_list}
\alias{create_competition_list}
\title{create_competition_list}
\usage{
create_competition_list(URL)
}
\arguments{
\item{URL}{URL for a competition}
}
\value{
a list of all events
}
\description{
create_competition_list
}
\examples{
 \dontrun{
   create_competition_list(url)
}
}
