% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_name_club_map}
\alias{data_name_club_map}
\title{Variable map for mapping names to club and country}
\format{
A list of the variables included to engineer dataset features
\describe{
\item{data_name_club_map}{name}
\item{data_name_club_map}{country}
\item{data_name_club_map}{name_other}
\item{data_name_club_map}{club}
}
}
\usage{
data(data_name_club_map)
}
\description{
List of names  - DK only
}
\keyword{datasets}
