% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_name_variation_map}
\alias{data_name_variation_map}
\title{Variable map for mapping name variations}
\format{
A list of the variables included to engineer dataset features
\describe{
\item{data_name_club_map}{name}
\item{data_name_club_map}{name_1}
}
}
\usage{
data(data_name_variation_map)
}
\description{
Name variations - KTK ONLY
}
\keyword{datasets}
