% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_representing_map}
\alias{data_representing_map}
\title{Variable map for mapping representing to country and club}
\format{
A list of the variables included to engineer dataset features
\describe{
\item{data_representing_map}{representing}
\item{data_representing_map}{country}
\item{data_representing_map}{club}
}
}
\usage{
data(data_representing_map)
}
\description{
List of clubs
}
\keyword{datasets}
