% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_event.R
\name{fetch_event}
\alias{fetch_event}
\title{Return the CSV file of one event}
\usage{
fetch_event(id)
}
\arguments{
\item{id}{id of an event}
}
\value{
tibble of a csv
}
\description{
Return the CSV file of one event
}
\examples{
 \dontrun{
   fetch_event(id)
}
}
