% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_past_event_list.R
\name{fetch_past_event_list}
\alias{fetch_past_event_list}
\title{create a list of held competitions from Sporttech.io}
\usage{
fetch_past_event_list()
}
\value{
a list of all events
}
\description{
create a list of held competitions from Sporttech.io
}
\examples{
 \dontrun{
   fetch_past_event_list()
}
}
