% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_event.R
\name{process_event}
\alias{process_event}
\title{Process and individual event}
\usage{
process_event(file)
}
\arguments{
\item{file}{path to a competition file}
}
\value{
a tibble of all scores
}
\description{
Process and individual event
}
\examples{
 \dontrun{
   process_event(file)
}
}
