% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_event_data.R
\name{read_all_events_data}
\alias{read_all_events_data}
\title{Read event data}
\usage{
read_all_events_data(URL)
}
\arguments{
\item{URL}{URL for a competition}
}
\value{
a list of all events
}
\description{
Read event data
}
\examples{
 \dontrun{
   read_all_events_data(url)
}
}
