% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_representing.R
\name{clean_representing}
\alias{clean_representing}
\title{Create a club and country based on representing column}
\usage{
clean_representing(representing, title, col)
}
\arguments{
\item{representing}{column from sporttech.  Can contain country, club or both country and club}

\item{title}{title of the event}

\item{col}{1 or 2.  1 for country and 2 for club}
}
\value{
dataset with representing, country and club
}
\description{
Create a club and country based on representing column
}
\examples{
 \dontrun{
   clean_representing(representing, title, col)
}
}
