% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_event_local.R
\name{fetch_event_local}
\alias{fetch_event_local}
\title{Return the CSV file of one event cleaned}
\usage{
fetch_event_local(file)
}
\arguments{
\item{file}{a local file in csv format from sporttech.io}
}
\value{
tibble of a csv
}
\description{
Return the CSV file of one event cleaned
}
\examples{
 \dontrun{
   fetch_event_local(file)
}
}
