% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_event_url.R
\name{fetch_event_url}
\alias{fetch_event_url}
\title{Return the CSV file of one event cleaned}
\usage{
fetch_event_url(id)
}
\arguments{
\item{id}{id of an event}
}
\value{
tibble of a csv
}
\description{
Return the CSV file of one event cleaned
}
\examples{
 \dontrun{
   fetch_event_url(id)
}
}
