% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_event.R
\name{process_event}
\alias{process_event}
\title{Process an individual event}
\usage{
process_event(event)
}
\arguments{
\item{event}{tibble of a competition}
}
\value{
a tibble of all scores for a single competition
}
\description{
Process an individual event
}
\examples{
 \dontrun{
   process_event(file)
}
}
