% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_event_list.R
\name{process_event_list}
\alias{process_event_list}
\title{Process_event_list}
\usage{
process_event_list(event_list)
}
\arguments{
\item{event_list}{list of all trampoline events}
}
\value{
a tibble of all trampoline events
}
\description{
Process_event_list
}
\examples{
 \dontrun{
   process_event_list(list)
}
}
